added grid in user management
